DERBY-6000: Add JDBC 4.2 large max rows methods to Derby's embedded and client implementations of java.sql.Statement.
git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1440656 13f79535-47bb-0310-9956-ffa450edef68

+++ java/client/org/apache/derby/client/am/Cursor.java
@@ -82,7 +82,7 @@ public abstract class Cursor {
 
     // Total number of rows read so far.
     // This should never exceed this.statement.maxRows
-    int rowsRead_;
+    long rowsRead_;
 
     // Maximum column size limit in bytes.
     int maxFieldSize_ = 0;
